[trigger] Add support for trigger on embed contents with toggle
Add toggle option to enable searching in embeds for triggers.
Changed re.findall to re.search as you don't need to return all occurrences, this way the search stops as soon as it finds the string instead of crawling the whole contents.
Moved the cooldown check outside of the search as all three search checks use it and there is no need for the string search if the cooldown hasn't been met.
A potential alternative is to move it out to it's own function.